requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.0
billiard==3.6.4.0
black==23.3.0
blis==0.7.9
Brotli==1.1.0
catalogue==2.0.8
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
confection==0.0.4
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2022.1
cymem==2.0.7
dataclasses-json==0.5.7
Deprecated==1.2.18
distlib==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
django-crispy-forms==2.0
django-filter==22.1
django-tenants==3.8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
exceptiongroup==1.2.2
filelock==3.9.0
frozenlist==1.3.3
fsspec==2023.5.0
gensim==4.3.1
googleapis-common-protos==1.59.1
gpt-index==0.6.16.post1
greenlet==2.0.2
grpcio==1.73.0
gunicorn==23.0.0
huggingface-hub==0.15.1
idna==3.4
importlib_metadata==8.7.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
kombu==5.2.4
langchain==0.0.186
langcodes==3.3.0
llama-index==0.6.16.post1
loguru==0.7.3
Markdown==3.4.1
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.24.3
openai==0.27.1
openapi-schema-pydantic==1.2.4
opentelemetry-api==1.34.1
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-dbapi==0.55b1
opentelemetry-instrumentation-django==0.55b1
opentelemetry-instrumentation-psycopg2==0.55b1
opentelemetry-instrumentation-requests==0.55b1
opentelemetry-instrumentation-wsgi==0.55b1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-util-http==0.55b1
ordered-set==4.1.0
packaging==23.0
pandas==2.0.2
pathspec==0.11.0
pathy==0.10.1
Pillow==10.0.1
pipenv==2023.2.18
platformdirs==3.1.0
pluggy==1.0.0
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==3.20.3
psycopg2-binary==2.9.6
pydantic==1.10.8
PyJWT==2.6.0
pytest==7.3.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2022.7.1
PyYAML==6.0
rake-nltk==1.0.6
regex==2023.5.5
requests==2.28.2
ruamel.yaml==0.17.22
ruamel.yaml.clib==0.2.12
ruff==0.1.7
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
soupsieve==2.4
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.15
sqlparse==0.4.3
srsly==2.4.6
tenacity==8.2.2
textblob==0.17.1
thinc==8.1.10
threadpoolctl==3.1.0
thrift==0.22.0
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.2.1
tqdm==4.65.0
transformers==4.29.2
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
virtualenv==20.20.0
virtualenv-clone==0.5.7
wasabi==1.1.2
wcwidth==0.2.6
whitenoise==6.9.0
wrapt==1.17.2
yarl==1.8.2
zipp==3.22.0
# autogen==0.2.10